fix(client/wasm): exclude gotreesitter/gob from the TinyGo WASM client
runtime.wasm was throwing an uncaught RuntimeError: unreachable on every
/admin/editor load in production. engine/surface/lowering.go and
ir/lower.go — both host/build-time-only CST-to-bytecode lowering code,
never called by the WASM client at runtime — had no build constraint
excluding them, so importing the ir package for its data types
transitively linked github.com/odvcencio/gotreesitter (and its
encoding/gob dependency) into every TinyGo build of client/wasm.

TinyGo's internal/reflectlite has a known gap: AssignableTo/Implements
against a non-empty interface panics ("reflect: unimplemented:
AssignableTo with interface"). gob's type-info construction
(mustGetTypeInfo -> userType -> implementsInterface) hits exactly this
during WASM boot, before any hydrate call — an unrecoverable panic that,
combined with the production build's -panic=trap flag, silently traps as
a bare unreachable with zero diagnostic output.

Tag both files !tinygo/!js (matching the existing exclusion pattern for
compile.go/grammar.go documented in compile_stub_tinygo.go) so the WASM
client never links the .gsx compiler's gotreesitter dependency it was
never meant to need. isValidGoIdent moves to its own unconstrained file
(ir/goident.go) since validate.go — always built — depends on it.

Adds a regression test asserting gotreesitter is absent from the TinyGo
client/wasm dependency closure.
